pyproject.toml
[tool.poetry]
name = "celerytest"
version = "0.1.0"
description = ""
authors = ["DaiZack <dzhengang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
celery = "^5.2.7"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
redis = "^4.5.4"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
